add unit tests for logs exporter
